Struct gmime::ParserOptions

pub struct ParserOptions(_);

A set of parser options used by Parser and various other parsing functions.

Methods

impl ParserOptions

pub fn new() -> ParserOptions

Creates a new set of ParserOptions.

Returns

a newly allocated set of ParserOptions with the default values.

pub fn clone(&mut self) -> Option<ParserOptions>

Clones a ParserOptions.

Returns

a newly allocated ParserOptions.

pub fn get_address_compliance_mode(&mut self) -> RfcComplianceMode

Gets the compliance mode that should be used when parsing rfc822 addresses.

Note: Even in RfcComplianceMode::Strict mode, the address parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

Returns

the compliance mode that is currently set.

pub fn get_allow_addresses_without_domain(&mut self) -> bool

Gets whether or not the rfc822 address parser should allow addresses without a domain.

In general, you'll probably want this value to be false (the default) as it allows maximum interoperability with existing (broken) mail clients and other mail software such as sloppily written perl scripts (aka spambots) that do not properly quote the name when it contains a comma.

This option exists in order to allow parsing of mailbox addresses that do not have a domain component. These types of addresses are rare and were typically only used when sending mail to other users on the same UNIX system.

Returns

true if the address parser should allow addresses without a domain.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_fallback_charsets(&mut self) -> Vec<String>

Gets the fallback charsets to try when decoding 8-bit headers.

Returns

a None-terminated list of charsets to try when decoding 8-bit headers.

pub fn get_parameter_compliance_mode(&mut self) -> RfcComplianceMode

Gets the compliance mode that should be used when parsing Content-Type and Content-Disposition parameters.

Note: Even in RfcComplianceMode::Strict mode, the parameter parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

Returns

the compliance mode that is currently set.

pub fn get_rfc2047_compliance_mode(&mut self) -> RfcComplianceMode

Gets the compliance mode that should be used when parsing rfc2047 encoded words.

Note: Even in RfcComplianceMode::Strict mode, the rfc2047 parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

Returns

the compliance mode that is currently set.

pub fn set_address_compliance_mode(&mut self, mode: RfcComplianceMode)

Sets the compliance mode that should be used when parsing rfc822 addresses.

In general, you'll probably want this value to be RfcComplianceMode::Loose (the default) as it allows maximum interoperability with existing (broken) mail clients and other mail software such as sloppily written perl scripts (aka spambots).

Note: Even in RfcComplianceMode::Strict mode, the address parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

mode

a RfcComplianceMode

pub fn set_allow_addresses_without_domain(&mut self, allow: bool)

Sets whether the rfc822 address parser should allow addresses without a domain.

In general, you'll probably want this value to be false (the default) as it allows maximum interoperability with existing (broken) mail clients and other mail software such as sloppily written perl scripts (aka spambots) that do not properly quote the name when it contains a comma.

This option exists in order to allow parsing of mailbox addresses that do not have a domain component. These types of addresses are rare and were typically only used when sending mail to other users on the same UNIX system.

allow

true if the parser should allow addresses without a domain or false otherwise

pub fn set_fallback_charsets(&mut self, charsets: &Vec<&str>)

Sets the fallback charsets to try when decoding 8-bit headers.

Note: It is recommended that the list of charsets start with utf-8 and end with iso-8859-1.

charsets

a None-terminated list of charsets or None for the default list

pub fn set_parameter_compliance_mode(&mut self, mode: RfcComplianceMode)

Sets the compliance mode that should be used when parsing Content-Type and Content-Disposition parameters.

In general, you'll probably want this value to be RfcComplianceMode::Loose (the default) as it allows maximum interoperability with existing (broken) mail clients and other mail software such as sloppily written perl scripts (aka spambots).

Note: Even in RfcComplianceMode::Strict mode, the parameter parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

mode

a RfcComplianceMode

pub fn set_rfc2047_compliance_mode(&mut self, mode: RfcComplianceMode)

Sets the compliance mode that should be used when parsing rfc2047 encoded words.

In general, you'll probably want this value to be RfcComplianceMode::Loose (the default) as it allows maximum interoperability with existing (broken) mail clients and other mail software such as sloppily written perl scripts (aka spambots).

Note: Even in RfcComplianceMode::Strict mode, the parameter parser is fairly liberal in what it accepts. Setting it to RfcComplianceMode::Loose just makes it try harder to deal with garbage input.

mode

a RfcComplianceMode

pub fn get_default() -> Option<ParserOptions>

Gets the default parser options.

Returns

the default parser options.