Struct gmime::PartIter

pub struct PartIter(_);

A MIME part iterator.

Methods

impl PartIter

pub fn new<P: IsA<Object>>(toplevel: &P) -> PartIter

Creates a new PartIter for iterating over toplevel's subparts.

toplevel

a Object to use as the toplevel

Returns

a newly allocated PartIter which should be freed using PartIter::free when finished with it.

pub fn clone(&mut self) -> Option<PartIter>

Clones the self, including its current state.

Returns

a new PartIter that is identical to self.

pub fn get_current(&mut self) -> Option<Object>

Gets the Object at the current PartIter position.

Returns

the current Object or None if the state of self is invalid.

pub fn get_parent(&mut self) -> Option<Object>

Gets the parent of the Object at the current PartIter position.

Returns

the parent Object or None if the state of self is invalid.

pub fn get_path(&mut self) -> Option<String>

Gets the path of the current Object in the MIME structure used to initialize self.

Returns

a newly allocated string representation of the path to the Object at the current PartIter position.

pub fn get_toplevel(&mut self) -> Option<Object>

Gets the toplevel Object used to initialize self.

Returns

the toplevel Object.

pub fn is_valid(&mut self) -> bool

Checks that the current state of self is valid.

Returns

true if self is valid or false otherwise.

pub fn jump_to(&mut self, path: &str) -> bool

Updates the state of self to point to the Object specified by path.

path

a string representing the path to jump to

Returns

true if the Object specified by path exists or false otherwise.

pub fn next(&mut self) -> bool

Advances to the next part in the MIME structure used to initialize self.

Returns

true if successful or false otherwise.

pub fn prev(&mut self) -> bool

Rewinds to the previous part in the MIME structure used to initialize self.

Returns

true if successful or false otherwise.

pub fn remove(&mut self) -> bool

Removes the Object at the current position from its parent. If successful, self is advanced to the next position (since the current position will become invalid).

Returns

true if the part at the current position was removed or false otherwise.

pub fn replace<P: IsA<Object>>(&mut self, replacement: &P) -> bool

Replaces the Object at the current position with replacement.

replacement

a Object

Returns

true if the part at the current position was replaced or false otherwise.

pub fn reset(&mut self)

Resets the state of self to its initial state.