Struct gmime::StreamFs

pub struct StreamFs(_, _);

A Stream wrapper around POSIX file descriptors.

Implements

StreamFsExt, StreamExt

Methods

impl StreamFs

pub fn new(fd: i32) -> StreamFs

Creates a new StreamFs object around fd.

fd

a file descriptor

Returns

a stream using fd.

pub fn new_with_bounds(fd: i32, start: i64, end: i64) -> StreamFs

Creates a new StreamFs object around fd with bounds start and end.

fd

a file descriptor

start

start boundary

end

end boundary

Returns

a stream using fd with bounds start and end.

pub fn open(path: &str, flags: i32, mode: i32) -> Result<Stream, Error>

Creates a new StreamFs object for the specified path.

path

the path to a file

flags

as in open(2)

mode

as in open(2)

Returns

a stream using for reading and/or writing to the specified file path or None on error.