Trait gmime::DecryptResultExt

pub trait DecryptResultExt {
    fn get_mdc(&self) -> DigestAlgo;

    fn get_recipients(&self) -> Option<CertificateList>;

    fn get_session_key(&self) -> Option<String>;

    fn get_signatures(&self) -> Option<SignatureList>;

    fn set_mdc(&self, mdc: DigestAlgo);

    fn set_recipients(&self, recipients: &CertificateList);

    fn set_session_key<'a, P: Into<Option<&'a str>>>(&self, session_key: P);

    fn set_signatures(&self, signatures: &SignatureList);
}

Trait containing all DecryptResult methods.

Implementors

DecryptResult

Required Methods

fn get_mdc(&self) -> DigestAlgo

Get the mdc digest algorithm used.

Returns

the mdc digest algorithm used.

fn get_recipients(&self) -> Option<CertificateList>

Gets the list of certificates that the stream had been encrypted to.

Returns

a CertificateList.

fn get_session_key(&self) -> Option<String>

Get the session key used for this decryption.

Returns

the session key digest algorithm used, or None if no session key was requested or found.

fn get_signatures(&self) -> Option<SignatureList>

Gets a list of signatures if the encrypted stream had also been signed.

Returns

a SignatureList or None if the stream was not signed.

fn set_mdc(&self, mdc: DigestAlgo)

Set the mdc digest algorithm used.

mdc

a DigestAlgo

fn set_recipients(&self, recipients: &CertificateList)

Sets the list of certificates that the stream had been encrypted to.

recipients

A CertificateList

fn set_session_key<'a, P: Into<Option<&'a str>>>(&self, session_key: P)

Set the session key to be returned by this decryption result.

session_key

a string representing the session key or None to unset the key

fn set_signatures(&self, signatures: &SignatureList)

Sets the list of signatures.

signatures

A SignatureList