Trait gmime::ParamListExt

pub trait ParamListExt {
    fn clear(&self);

    fn get_parameter(&self, name: &str) -> Option<Param>;

    fn get_parameter_at(&self, index: i32) -> Option<Param>;

    fn length(&self) -> i32;

    fn remove(&self, name: &str) -> bool;

    fn remove_at(&self, index: i32) -> bool;

    fn set_parameter(&self, name: &str, value: &str);
}

Trait containing all ParamList methods.

Implementors

ParamList

Required Methods

fn clear(&self)

Clears the list of parameters.

fn get_parameter(&self, name: &str) -> Option<Param>

Gets the Param with the given name.

name

the name of the parameter

Returns

the requested Param.

fn get_parameter_at(&self, index: i32) -> Option<Param>

Gets the Param at the specified index.

index

the index of the requested parameter

Returns

the Param at the specified index.

fn length(&self) -> i32

Gets the length of the list.

Returns

the number of Param items in the list.

fn remove(&self, name: &str) -> bool

Removes a parameter from the ParamList.

name

the name of the parameter

Returns

true if the specified parameter was removed or false otherwise.

fn remove_at(&self, index: i32) -> bool

Removes a Param from the ParamList at the specified index.

index

index of the param to remove

Returns

true if a Param was removed or false otherwise.

fn set_parameter(&self, name: &str, value: &str)

Sets the specified parameter to value.

name

The name of the parameter

value

The parameter value