Trait gmime::SignatureListExt

pub trait SignatureListExt {
    fn add(&self, sig: &Signature) -> i32;

    fn clear(&self);

    fn contains(&self, sig: &Signature) -> bool;

    fn get_signature(&self, index: i32) -> Option<Signature>;

    fn index_of(&self, sig: &Signature) -> i32;

    fn insert(&self, index: i32, sig: &Signature);

    fn length(&self) -> i32;

    fn remove(&self, sig: &Signature) -> bool;

    fn remove_at(&self, index: i32) -> bool;

    fn set_signature(&self, index: i32, sig: &Signature);
}

Trait containing all SignatureList methods.

Implementors

SignatureList

Required Methods

fn add(&self, sig: &Signature) -> i32

Adds a Signature to the SignatureList.

sig

a Signature

Returns

the index of the added Signature.

fn clear(&self)

Clears the list of addresses.

fn contains(&self, sig: &Signature) -> bool

Checks whether or not the specified Signature is contained within the SignatureList.

sig

a Signature

Returns

true if the specified Signature is contained within the specified SignatureList or false otherwise.

fn get_signature(&self, index: i32) -> Option<Signature>

Gets the Signature at the specified index.

index

index of Signature to get

Returns

the Signature at the specified index or None if the index is out of range.

fn index_of(&self, sig: &Signature) -> i32

Gets the index of the specified Signature inside the SignatureList.

sig

a Signature

Returns

the index of the requested Signature within the SignatureList or %-1 if it is not contained within the SignatureList.

fn insert(&self, index: i32, sig: &Signature)

Inserts a Signature into the SignatureList at the specified index.

index

index to insert at

sig

a Signature

fn length(&self) -> i32

Gets the length of the list.

Returns

the number of Signature objects in the list.

fn remove(&self, sig: &Signature) -> bool

Removes a Signature from the SignatureList.

sig

a Signature

Returns

true if the specified Signature was removed or false otherwise.

fn remove_at(&self, index: i32) -> bool

Removes a Signature from the SignatureList at the specified index.

index

index to remove

Returns

true if an Signature was removed or false otherwise.

fn set_signature(&self, index: i32, sig: &Signature)

Sets the Signature at the specified index to sig.

index

index of Signature to set

sig

a Signature