Crate gmime

Structs

ApplicationPkcs7Mime

AutocryptHeader

AutocryptHeaderList

Bytes
A shared immutable byte slice (the equivalent of Rc<[u8]>).

Certificate

CertificateList

ContentDisposition

ContentType

CryptoContext

DataWrapper

DateTime

DecryptFlags

DecryptResult

Error
A generic error capable of representing various error domains (types).

File

Filter

FilterBasic

FilterBest

FilterBestFlags

FilterCharset

FilterDos2Unix

FilterEnriched

FilterFrom

FilterGZip

FilterHTML

FilterSmtpData

FilterStrip

FilterUnix2Dos

FilterWindows

FilterYenc

FormatOptions

GpgContext

Header

HeaderList

IOCondition

InternetAddress

InternetAddressList

InternetAddressMailbox

Message

MessagePart

MessagePartial

Multipart

MultipartEncrypted

MultipartSigned

Object

Param

ParamList

Parser

ParserOptions

Part

PartIter

Pkcs7Context

Priority
The priority of sources

Signature

SignatureList

Source

Stream

StreamBuffer

StreamCat

StreamFile

StreamFilter

StreamFs

StreamGIO

StreamMem

StreamMmap

StreamNull

StreamPipe

TextPart

Variant
A generic immutable value capable of carrying various types.

VariantType
Describes Variant types.

Enums

AddressType

AutocryptPreferEncrypt

ChecksumType

ContentEncoding

DigestAlgo

EncodingConstraint

EncryptFlags

FilterFromMode

FilterGZipMode

Format

NewLineFormat

OpenPGPData

ParamEncodingMethod

PubKeyAlgo

RfcComplianceMode

SecureMimeType

SeekType

SeekWhence

SignatureStatus

StreamBufferMode

Trust

Validity

VerifyFlags

Traits

ApplicationPkcs7MimeExt

AutocryptHeaderExt

AutocryptHeaderListExt

CertificateExt

CertificateListExt

ContentDispositionExt

ContentTypeExt

CryptoContextExt

DataWrapperExt

DecryptResultExt

FilterBestExt

FilterExt

FilterGZipExt

FilterWindowsExt

FilterYencExt

HeaderExt

HeaderListExt

InternetAddressExt

InternetAddressListExt

InternetAddressMailboxExt

MessageExt

MessagePartExt

MessagePartialExt

MultipartEncryptedExt

MultipartExt

MultipartSignedExt

ObjectExt

ParamExt

ParamListExt

ParserExt

PartExt

SignatureExt

SignatureListExt

StreamCatExt

StreamExt

StreamFileExt

StreamFilterExt

StreamFsExt

StreamGIOExt

StreamMemExt

StreamMmapExt

StreamNullExt

StreamPipeExt

TextPartExt